refactor(getOrderedToolbarItems): moved to utils file
